fixed annoying implicit conversion warnings
svn path=/trunk/iCub/; revision=5056

+++ src/firmware/controllerBll/Code/filters.c
@@ -12,7 +12,7 @@ Int32 lpf_ord1_3hz(Int32 input, int j)
         y_filt[0][j] = y_filt[1][j]; 
         y_filt[1][j] = (x_filt[0][j] + x_filt[1][j])
         	 	  + (  0.9813258905  * y_filt[0][j]);
-        return y_filt[1][j];
+        return (Int32)(y_filt[1][j]);
 }
 
 Int32 lpf_ord1_30hz(Int32 input, int j)
@@ -24,7 +24,7 @@ Int32 lpf_ord1_30hz(Int32 input, int j)
         y_filt[0][j] = y_filt[1][j]; 
         y_filt[1][j] =   (x_filt[0][j] + x_filt[1][j])
                       + (  0.8272719460  * y_filt[0][j]);
-        return y_filt[1][j];
+        return (Int32)(y_filt[1][j]);
 }
 
          
@@ -37,7 +37,7 @@ Int32 lpf_ord2_30hz(Int32 input, int j)
         y_filt[0][j] = y_filt[1][j]; y_filt[1][j] = y_filt[2][j]; 
         y_filt[2][j] =   (x_filt[0][j] + x_filt[2][j]) + 2 * x_filt[1][j]
                      + ( -0.7660066009 * y_filt[0][j]) + (  1.7347257688 * y_filt[1][j]);
-        return y_filt[2][j];
+        return (Int32)(y_filt[2][j]);
 }
    
 
@@ -52,7 +52,7 @@ Int32 lpf_ord4_30hz(Int32 input, int j)
         y_filt[4][j] =   (x_filt[0][j] + x_filt[4][j]) + 4 * (x_filt[1][j] + x_filt[3][j]) + 6 * x_filt[2][j]
                      + ( -0.6105348076 * y_filt[0][j]) + (  2.7426528211 * y_filt[1][j])
                      + ( -4.6409024127 * y_filt[2][j]) + (  3.5077862074 * y_filt[3][j]);
-        return y_filt[4][j];	
+        return (Int32)(y_filt[4][j]);	
 }
